refactor: optional get 함수 대신 orElseThrow 함수로 변경

+++ src/main/java/techcourse/myblog/controller/CommentController.java
@@ -8,8 +8,6 @@ import org.springframework.web.bind.annotation.*;
 import techcourse.myblog.controller.dto.CommentDto;
 import techcourse.myblog.domain.*;
 
-import javax.servlet.http.HttpSession;
-
 import static techcourse.myblog.controller.CommentController.COMMENT_URL;
 
 @Slf4j
@@ -24,7 +22,7 @@ public class CommentController {
     @PostMapping
     public String save(CommentDto commentDto, User user) {
         log.debug(">>> post commentDto : {}, user : {}", commentDto, user);
-        Article article = articleRepository.findById(commentDto.getArticleId()).get();
+        Article article = articleRepository.findById(commentDto.getArticleId()).orElseThrow(RuntimeException::new);
         Comment comment = commentDto.toComment(user, article);
         commentRepository.save(comment);
         return "redirect:/articles/" + commentDto.getArticleId();
@@ -33,15 +31,20 @@ public class CommentController {
     @DeleteMapping
     public String delete(CommentDto commentDto, User user) {
         log.debug(">>> delete commentDto : {}", commentDto);
-        Article article = articleRepository.findById(commentDto.getArticleId()).get();
-        Comment comment = commentDto.toComment(user, article);
+        Comment comment = commentRepository.findById(commentDto.getId()).orElseThrow(RuntimeException::new);
+        if (user.isNotMatch(comment.getUser())) {
+            return "redirect:/";
+        }
         commentRepository.delete(comment);
         return "redirect:/articles/" + commentDto.getArticleId();
     }
 
     @GetMapping("{id}/edit")
-    public String showEditPage(@PathVariable long id, Model model) {
-        Comment comment = commentRepository.findById(id).get();
+    public String showEditPage(@PathVariable long id, Model model, User user) {
+        Comment comment = commentRepository.findById(id).orElseThrow(RuntimeException::new);
+        if (user.isNotMatch(comment.getUser())) {
+            return "redirect:/";
+        }
         model.addAttribute("comment", comment);
         return "comment-edit";
     }
@@ -49,7 +52,7 @@ public class CommentController {
     @PutMapping
     public String update(CommentDto commentDto, User user) {
         log.debug(">>> update commentDto : {}", commentDto);
-        Article article = articleRepository.findById(commentDto.getArticleId()).get();
+        Article article = articleRepository.findById(commentDto.getArticleId()).orElseThrow(RuntimeException::new);
         Comment comment = commentDto.toComment(user, article);
         commentRepository.save(comment);
         return "redirect:/articles/" + article.getId();
